requirements.txt
absl-py==0.12.0
alembic==1.0.11
amqp==5.0.6
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
d2l==0.16.6
dataclasses==0.6
debtcollector==2.2.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
dm-tree==0.1.6
dnspython==1.16.0
dogpile.cache==1.1.3
entrypoints==0.3
eventlet==0.31.0
extras==1.0.0
fasteners==0.16.3
fixtures==3.0.0
flatbuffers==1.12
future==0.18.2
futurist==2.3.0
gast==0.4.0
gin-config==0.4.0
gluon==1.1.0
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.20.0
google-cloud-core==1.7.1
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.19.1
idna==2.10
importlib-resources==5.2.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.14
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keystoneauth1==4.3.1
keystonemiddleware==9.3.0
kiwisolver==1.3.1
kombu==5.1.0
linecache2==1.0.0
logutils==0.3.5
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netaddr==0.8.0
netifaces==0.11.0
notebook==6.4.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python-headless==4.5.2.54
opt-einsum==3.3.0
os-service-types==1.7.0
oslo.cache==2.8.1
oslo.concurrency==4.4.0
oslo.config==8.7.0
oslo.context==3.3.0
oslo.db==7.0.0
oslo.i18n==5.0.1
oslo.log==4.5.0
oslo.messaging==12.8.0
oslo.middleware==4.2.0
oslo.policy==3.8.1
oslo.serialization==4.1.0
oslo.service==2.6.0
oslo.utils==4.9.1
oslo.versionedobjects==2.4.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
Paste==3.5.0
PasteDeploy==2.1.1
pbr==5.6.0
pecan==1.4.0
pickleshare==0.7.5
Pillow==8.2.0
portalocker==2.0.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
proto-plus==1.19.0
protobuf==3.17.3
psutil==5.8.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycadf==3.1.1
pycocotools==2.0.2
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-etcd==0.4.5
python-keystoneclient==4.2.0
python-mimeparse==1.6.0
python-slugify==5.0.2
pytz==2021.1
pywin32==301
pywinpty==1.1.2
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
repoze.lru==0.7
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
Routes==2.5.1
rsa==4.7.2
sacrebleu==1.5.1
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
sentencepiece==0.1.96
seqeval==1.2.2
simplegeneric==0.8.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.0.19
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
statsd==3.3.0
stevedore==3.3.0
Tempita==0.5.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.1.0
tensorflow-model-optimization==0.6.0
tensorflow-text==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.4.0
text-unidecode==1.3
tf-models-official==2.4.0
tf-slim==1.1.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.61.1
traceback2==1.4.0
traitlets==5.0.5
typeguard==2.12.1
typing-extensions==3.7.4.3
unittest2==1.1.0
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
WebTest==2.0.35
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WSME==0.10.1
yappi==1.3.2
zipp==3.5.0
